% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{decode_sentiment}
\alias{decode_sentiment}
\title{Decode: Sentiment Word -> Numeric}
\usage{
decode_sentiment(words, lexicon)
}
\arguments{
\item{words}{Vector of words}

\item{lexicon}{The ordered character vector}
}
\description{
Decode: Sentiment Word -> Numeric
}
