% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{encode_sentiment}
\alias{encode_sentiment}
\title{Encode: Numeric -> Sentiment Word}
\usage{
encode_sentiment(x, lexicon)
}
\arguments{
\item{x}{Numeric vector (standardized)}

\item{lexicon}{The ordered character vector}
}
\description{
Encode: Numeric -> Sentiment Word
}
