% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tense-package.R
\docType{package}
\name{tense-package}
\alias{tense}
\alias{tense-package}
\title{tense: tense is a LLM semantic sequence simulation technique for ordered numbers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

tense is an text-based LLM semantic expression implementation for ordered numeric sequences. Using LLMs to predict the next token, tense is time-based estimation via semantic sentiment encoding & quantization for LLMs.
}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@proton.me}

}
\keyword{internal}
