% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tense_init}
\alias{tense_init}
\title{Initialize Tense Model}
\usage{
tense_init(model_name = "llama3", temperature = 0.9, context_window = 40)
}
\arguments{
\item{model_name}{LLM model name.}

\item{temperature}{a temperature for the LLM}

\item{context_window}{Steps to look back.}
}
\description{
Initialize Tense Model
}
