% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tense-package.R
\docType{package}
\name{tense-package}
\alias{tense}
\alias{tense-package}
\title{tense: TENSE: a time-llm model}
\description{
tense is an text-based time-LLM semantic expression implementation. Using LLMs to predict the next token, tense is time-based estimation via semantic sentiment encoding & quantization for LLMs.
}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@proton.me}

}
\keyword{internal}
